+++ ziptastic-client/src/Ziptastic/Client.hs
@@ -1,8 +1,4 @@
-{-# LANGUAGE
-    DataKinds
-  , TypeOperators
-  , OverloadedStrings
-#-}
+{-# LANGUAGE DataKinds, TypeOperators, OverloadedStrings #-}
 
 -- | This module provides a simple interface to Ziptastic's forward and reverse geocoding API
 -- (<https://www.getziptastic.com/>).
@@ -33,26 +29,31 @@ module Ziptastic.Client
 import           Data.ISO3166_CountryCodes (CountryCode)
 import           Data.Proxy (Proxy(..))
 import           Data.Text (Text)
-import           Network.HTTP.Client (Manager)
 import           Servant.API ((:<|>)(..))
 import           Servant.Client
-  ( BaseUrl(..), ClientEnv(..), ClientM, Scheme(..), ServantError
+  ( BaseUrl(..), ClientEnv(..), ClientM, Scheme(..), ServantError(..)
   , client, runClientM
   )
 
 import           Ziptastic.Core (ApiKey, ForApi(..), LocaleInfo)
 import qualified Ziptastic.Core as Core
 
+-- TODO: Needed for hack (see below).
+import           Network.HTTP.Client
+  ( Manager, HttpException(HttpExceptionRequest), HttpExceptionContent(TlsNotSupported) )
+import           Control.Exception (fromException)
+
 -- | Performs a forward geocode lookup at the given country and postal code.
 --
 -- The success result is a list because in rare cases you may receive multiple records.
 -- If the request fails the result will be 'Left' with an error.
 forwardGeocode :: ApiKey
-               -> Manager      -- ^ HTTP connection manager
+               -> Manager      -- ^ HTTP connection manager (if TLS is supported, request will be made over HTTPS)
                -> CountryCode  -- ^ country
                -> Text         -- ^ postal code
                -> IO (Either ServantError [LocaleInfo])
-forwardGeocode apiKey manager countryCode postalCode = runClientM func (ClientEnv manager baseUrl)
+forwardGeocode apiKey manager countryCode postalCode =
+  tryTlsOrDegrade $ \scheme -> runClientM func (ClientEnv manager $ baseUrl scheme)
   where func = forwardGeocode' (apiClient apiKey) (ForApi countryCode) postalCode
 
 -- | Performs a reverse geocode lookup at the given coordinates using a default radius of 5000 meters.
@@ -60,7 +61,7 @@ forwardGeocode apiKey manager countryCode postalCode = runClientM func (ClientEn
 -- The success result is a list because in rare cases you may receive multiple records.
 -- If the request fails the result will be 'Left' with an error.
 reverseGeocode :: ApiKey
-               -> Manager -- ^ HTTP connection manager
+               -> Manager -- ^ HTTP connection manager (if TLS is supported, request will be made over HTTPS)
                -> Double  -- ^ latitude
                -> Double  -- ^ longitude
                -> IO (Either ServantError [LocaleInfo])
@@ -71,12 +72,13 @@ reverseGeocode apiKey manager lat long = reverseGeocodeWithRadius apiKey manager
 -- The success result is a list because in rare cases you may receive multiple records.
 -- If the request fails the result will be 'Left' with an error.
 reverseGeocodeWithRadius :: ApiKey
-                         -> Manager -- ^ HTTP connection manager
+                         -> Manager -- ^ HTTP connection manager (if TLS is supported, request will be made over HTTPS)
                          -> Double  -- ^ latitude
                          -> Double  -- ^ longitude
                          -> Int     -- ^ radius (in meters)
                          -> IO (Either ServantError [LocaleInfo])
-reverseGeocodeWithRadius apiKey manager lat long radius = runClientM func (ClientEnv manager baseUrl)
+reverseGeocodeWithRadius apiKey manager lat long radius =
+  tryTlsOrDegrade $ \scheme -> runClientM func (ClientEnv manager $ baseUrl scheme)
   where func = reverseGeocodeWithRadius' (mkReverseGeocode (apiClient apiKey) lat long) radius
 
 
@@ -104,11 +106,26 @@ apiClient apiKey = ApiClient
       where
         withRadius :<|> withDefaultRadius = reverseGeocodeApi lat long
 
-baseUrl :: BaseUrl
-baseUrl = BaseUrl
-  { baseUrlScheme = if Core.baseUrlIsHttps then Https else Http
+baseUrl :: Scheme -> BaseUrl
+baseUrl scheme = BaseUrl
+  { baseUrlScheme = scheme
   , baseUrlHost   = Core.baseUrlHost
-  , baseUrlPort   = Core.baseUrlPort
+  , baseUrlPort   = case scheme of
+      Http  -> 80
+      Https -> 443
   , baseUrlPath   = Core.baseUrlPath
   }
 
+
+-- TODO: Hack to detect if a Manager supports TLS.
+-- See https://github.com/snoyberg/http-client/issues/266
+tryTlsOrDegrade :: (Scheme -> IO (Either ServantError a)) -> IO (Either ServantError a)
+tryTlsOrDegrade req = do
+  secureResponse <- req Https
+  case secureResponse of
+    Left (ConnectionError connE) -> case fromException connE of
+      Just (ConnectionError connE2) -> case fromException connE2 of
+        Just (HttpExceptionRequest _ TlsNotSupported) -> req Http
+        _ -> pure secureResponse
+      _ -> pure secureResponse
+    _ -> pure secureResponse