+++ ziptastic-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0.0
+
+  * Detect if a `Manager` supports TLS and switch between HTTP and HTTPS automatically.
+  * Change LANGUAGE pragma formatting to support older GHCs.
+
 ## 0.2.0.0
 
   * For failures, provide full `ServantError` instead converting it to `String`.