+++ ziptastic-core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0.0
+
+  * Don't export scheme and port for base URL since Ziptastic supports both HTTP and HTTPS.
+  * Change LANGUAGE pragma formatting to support older GHCs.
+
 ## 0.1.0.1
 
   * Add version bounds.